Disable Transparent Huge Pages in the kernel cmdline
Originally I wanted to fix the permissions for that systemd
unit until I realized what it does. It's a straight c&p from
a mongodb manual that suggests that THP might be bad for
sparse reads (why not at least use madvise here instead
of never?!).

Anyway, as I do not want to change runtime behavior now,
this commit for now only moves the command to grubs kernel
cmdline. Cleanly seperated as a new config file so our
concerns are neatly seperated.