Add support for the unified gUM cluster frontend option
This makes shure all nodes appear as a single
host to the user, with the browser only prompting
for camera and microphone permission once for
the entire cluster.

By defining

  bbb_cluster_proxy_host="bbb.yourdomain.example"

this change will allow hiding all nodes behind a
single hostname, such as bbb.yourdomain.example.
The Nodes will appear in the URL.

This requires a web server listening on

  bbb.yourdomain.example

with a redirect as follows:

location ~ "^/(\d{4})/html5client/(.*)$" {
  resolver 8.8.8.8; # Use something local for prod!
  proxy_pass https://$1.bbb.yourdomain.example;
  proxy_http_version 1.1;
  proxy_set_header Upgrade $http_upgrade;
  proxy_set_header Connection "Upgrade";
}

Note: This assumes a numbering scheme. You can also
template this out with one location per backend. Take
care not to open the regex too wide to avoid security
issues!